% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{rclinvarbitration_init}
\alias{rclinvarbitration_init}
\title{Initialize a ClinVar semantic graph}
\usage{
rclinvarbitration_init(con)
}
\arguments{
\item{con}{A DuckDB DBI connection.}
}
\value{
\code{con}, invisibly.
}
\description{
Initialize a ClinVar semantic graph
}
