% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RClinVarbitration-package}
\alias{RClinVarbitration}
\alias{RClinVarbitration-package}
\title{RClinVarbitration: relational ClinVar evidence in DuckDB}
\description{
A DuckDB-native, XML-first ClinVar store. Its package-owned native extension
streams official \code{.xml.gz} VCV releases through libxml2 in one pass. SQL
materializes VCV, allele, location, gene, RCV, SCV, condition, observation,
citation, attribute, and evidence-text relations; versioned
ClinVarbitration policy is deliberately derived from those source records.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sounkou-bioinfo/RClinVarbitration}
  \item Report bugs at \url{https://github.com/sounkou-bioinfo/RClinVarbitration/issues}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}

Authors:
\itemize{
  \item Sounkou Mahamane Toure \email{sounkoutoure@gmail.com}
}

Other contributors:
\itemize{
  \item DuckDB Foundation (Bundled DuckDB C extension headers) [copyright holder]
}

}
\keyword{internal}
