% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{rclinvarbitration_init}
\alias{rclinvarbitration_init}
\title{Initialize the ClinVar relational schema}
\usage{
rclinvarbitration_init(con)
}
\arguments{
\item{con}{A DuckDB DBI connection.}
}
\value{
\code{con}, invisibly.
}
\description{
Initialize the ClinVar relational schema
}
